+++ src/main/java/com/vincent/forexledger/model/entry/Entry.java
@@ -11,7 +11,7 @@ public class Entry {
     private TransactionType transactionType;
     private Date transactionDate;
     private double foreignAmount;
-    private Double twdAmount;
+    private Integer twdAmount;
     private String anotherBookId;
     private String creator;
     private Date createdTime;
@@ -56,11 +56,11 @@ public class Entry {
         this.foreignAmount = foreignAmount;
     }
 
-    public Double getTwdAmount() {
+    public Integer getTwdAmount() {
         return twdAmount;
     }
 
-    public void setTwdAmount(Double twdAmount) {
+    public void setTwdAmount(Integer twdAmount) {
         this.twdAmount = twdAmount;
     }
